server/requirements.txt
amqp==5.3.1
aniso8601==10.0.0
argh==0.31.3
attrs==25.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
distlib==0.3.9
dnspython==2.7.0
eventlet==0.39.1
filelock==3.18.0
Flask==2.3.3
flask-cors==5.0.1
Flask-SocketIO==5.3.6
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2
minknow_api==6.2.1
numpy==1.26.4
packaging==24.2
prompt_toolkit==3.0.50
protobuf==4.25.6
pycparser==2.22
pyRFC3339==1.1
pysam==0.23.0
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
redis==5.2.1
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
tzdata==2025.2
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
load-dotenv==0.1.0